Add files via upload
A Java program that prompts the user to enter a number and determines whether the number is positive or negative. If the number is positive, the program should print "The number is positive." If the number is negative, the program should print "The number is negative." and "the number is zero " in case of zero.